# template-utils-service-api/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  description: |
    "API for listing available templates, document structures and generating impersonated documents with provided data.
    The API has two operation groups: 
      - templates for generating single documents and 
      - documentstructures for generating document sets.
    The service is base on a template repository, a document structure repository and a result store. 
    These can be freely configured but once configured, have the same setup has to be used 
    for the template and document structure repositories since they have to be consitent. 
    
    The document generation is asynchronous, when posting a job you will get a transcation id (or provide by yourself) and 
    you can query and download the results based on that."  
  title: Document generation API
  version: 1.0.1
servers:
  - url: https://localhost:8080/api/v1
...
paths:
  /template:
    get:
      tags:
        - templates
      operationId: "getTemplates"
      summary: "Lists a page of the available document templates, or a given template in the template repository."
      x-contentType: application/json
      x-accepts: application/json
      parameters:
        - name: templateId
          description: "The template id as stored in the template repository."
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TemplateElementId"
        - $ref: "#/components/parameters/pageable"
      responses:
        "200":
          description: "Request processed successfully and data is present."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetTemplatesResponse"
        "400":
          description: Invalid DTO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "401":
          description: Bad credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
  /template/fill:
    post:
      tags:
        - templates
      operationId: "postTemplateGenerationJob"
      summary: "Posts a template generation with the given data"
      x-contentType: application/json
      x-accepts: application/json
      parameters:
        - name: templateId
          description: "The template id in the template repository. (e.g. for filesystem repo the relative path from the basedir - see get /templates)"
          explode: true
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/TemplateElementId"
        - name: notificationUrl
          description: "Optional notification webhook to indicate the end of the generation job."
          in: query
          required: false
          schema:
            type: string
            minLength: 0
            maxLength: 2048
            format: url
      requestBody:
        description: data provided for generation
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TemplateContext"
      responses:
        "202":
          description: Default
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateJobApiResponse"
        default:
          description: Default
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateJobApiResponse"
        "400":
          description: Invalid DTO
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "401":
          description: Bad credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "404":
          description: Template not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
  /template/fill/{transactionId}:
    get:
      tags:
        - templates
      operationId: "getResultDocumentByTransactionId"
      summary: "Returns the result document for a given transaction id and result document."
      x-contentType: application/json
      x-accepts: application/json
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/TransactionId"
      responses:
        "200":
          description: Default
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationResult"
        default:
          description: Default
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Bad credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
  /template/fill/{transactionId}/doc:
    get:
      tags:
        - templates
      operationId: "getResultDocumentForTemplateByTransactionIdAndResultDocumentId"
      summary: "Returns the result document for a given transaction id and result document."
      x-contentType: application/json
      x-accepts: application/json
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/TransactionId"
        - name: resultDocumentId
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/ResultDocumentId"
      responses:
        "200":
          description: Default
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        default:
          description: Default
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Bad credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
...
components:
  schemas:
...
    EmptyResponse:
      type: object
      description: "Generic empty response."
      additionalProperties: false
...
    Pageable:
      type: object
      description: "Pageable query parameters"
      additionalProperties: false
      properties:
        page:
          type: integer
          description: "The page to be returned."
          format: int32
          minimum: 0
          maximum: 32768
          default: 0
          example: 0
        size:
          type: integer
          description: "The number of items to be returned."
          format: int32
          minimum: 0
          maximum: 32768
          default: 10
          example: 10
        paged:
          type: boolean
          description: "Returns whether the current Pageable contains pagination information."
          default: true
        sort:
          type: array
          description: "Sort orders in the format fieldname,asc|desc. You may define multiple sort criteria."
          items:
            type: string
            format: "^([a-zA-Z0-9_-]*,asc|desc)*$"
            minLength: 0
            maxLength: 4096
            pattern: "^([a-zA-Z0-9_-]*,asc|desc)*$"
          minItems: 0
          maxItems: 64
    Page:
      type: object
      description: "A page is a sublist of a list of objects. It allows gain information about the position of it in the containing entire list."
      properties:
        totalPages:
          type: integer
          description: "Returns the number of total pages."
          format: uint32
          minimum: 0
          maximum: 32768
          default: 0
          example: 2
        totalElements:
          type: integer
          description: "Returns the total amount of elements."
          format: int64
          minimum: 0
          default: 0
          example: 18
          maximum: 2147483647
        number:
          type: integer
          description: "Returns the number of the current Slice. Is always non-negative."
          format: uint32
          minimum: 0
          maximum: 32768
          default: 0
          example: 1
        size:
          type: integer
          description: "Returns the number of elements currently on this Slice."
          format: uint32
          minimum: 0
          default: 8
          maximum: 32768
        numberOfElements:
          type: integer
          description: "Returns the number of elements currently on this Slice."
          format: uint32
          minimum: 0
          maximum: 32768
          default: 0
          example: 8

    # API specific types.
    # -------------------
...
    TemplateElementId:
      description: "The template id as it is stored in the template repository."
      type: string
      minLength: 0
      maxLength: 4000
      pattern: "^[a-zA-Z0-9_\\/-]*$"
      example: "integrationtests/contracts/contract_v09_hu.docx"
    TransactionId:
      type: string
      minLength: 0
      maxLength: 4000
      pattern: "^[a-zA-Z0-9_\\/-]*$"
    TemplateDocument:
      type: object
      additionalProperties: false
      properties:
        templateName:
          type: string
          minLength: 0
          maxLength: 4000
          pattern: "^[a-zA-Z0-9_/\\]*$"
        format:
          type: string
          minLength: 0
          maxLength: 50
          pattern: "^[a-zA-Z0-9_/\\]*$"
          enum: [DOCX, XLSX]
        locale:
          type: string
          minLength: 0
          maxLength: 50
          pattern: "^[a-zA-Z0-9_-/]*$"
        version:
          type: string
          minLength: 0
          maxLength: 4000
          pattern: "^[a-zA-Z0-9_/\\]*$"
        internalKey:
          type: string
          minLength: 0
          maxLength: 4000
          pattern: "^[a-zA-Z0-9_/\\]*$"
...
    DocumentStructureId:
      type: string
      minLength: 0
      maxLength: 4000
      pattern: "^[a-zA-Z0-9_\\/-]*$"
      example: 050bca79-5aba-4e32-a34d-9409edcb0a68
...
    ResultDocument:
      type: object
      additionalProperties: false
      properties:
        documentName:
          type: string
          minLength: 0
          maxLength: 4000
          pattern: "^[a-zA-Z0-9_/\\]*$"
        format:
          type: string
          minLength: 0
          maxLength: 50
          pattern: "^[a-zA-Z0-9_/\\]*$"
          enum: [DOCX, XLSX]
        url:
          type: string
          minLength: 0
          maxLength: 2048
          format: url
    GenerationResult:
      type: object
      additionalProperties: false
      properties:
        transactionId:
          $ref: "#/components/schemas/TransactionId"
        documentStructureId:
          $ref: "#/components/schemas/DocumentStructureId"
        elements:
          items:
            $ref: "#/components/schemas/ResultDocument"
          type: array
          maxItems: 32767
    ResultDocumentId:
      type: string
      minLength: 0
      maxLength: 2048
      pattern: "^[a-zA-Z0-9_/\\]*$"
    TemplateContext:
      type: object
      example: "
        {
        'ctx': {
        'org': {
        'orgCode': 'PB',
        'name': 'Vintage Services - Palm beach',
        'address': {
        'zip': 'Y-1234567',
        'city': 'Simply City',
        'address': 'Main blvd 432'
        }
        },
        'officer': {
        'name': 'Chuck Norris',
        'orgCode': 'PB-001',
        'login': 'PB\\cnorris'
        },
        'contract': {
        'contractor': {
        'name': 'John Doe',
        'birthDate': {
        'year': 1970,
        'month': 7,
        'day': 20
        }
        },
        'contractType': {
        'contractTypeName': 'Vintage Gold',
        'fee': 1500,
        'paymentFrequency': 'MONTHLY'
        },
        'beneficiaries': [
        {
        'beneficiaries': [
        {
        'phoneNumber': '+1 800 2234 567',
        'name': 'Jim Doe',
        'birthDate': {
        'year': 1975,
        'month': 8,
        'day': 11
        }
        },
        {
        'phoneNumber': '+1 800 2234 568',
        'name': 'Tim Doe',
        'birthDate': {
        'year': 1976,
        'month': 8,
        'day': 12
        }
        },
        {
        'phoneNumber': '+1 800 2234 569',
        'name': 'Pim Doe',
        'birthDate': {
        'year': 1977,
        'month': 8,
        'day': 13
        }
        },
        {
        'phoneNumber': '+1 800 3234 567',
        'name': 'Jack Ryan',
        'birthDate': {
        'year': 1962,
        'month': 8,
        'day': 11
        }
        },
        {
        'phoneNumber': '+1 800 3234 568',
        'name': 'John Goodall',
        'birthDate': {
        'year': 1946,
        'month': 8,
        'day': 11
        }
        },
        {
        'phoneNumber': '+1 800 3234 569',
        'name': 'Mortimer Young',
        'birthDate': {
        'year': 1991,
        'month': 8,
        'day': 11
        }
        },
        {
        'phoneNumber': '+1 800 3234 560',
        'name': 'Zack Black',
        'birthDate': {
        'year': 1987,
        'month': 8,
        'day': 11
        }
        }
        ],
        'phoneNumber': '+1 800 1234 567',
        'name': 'Jane Doe',
        'birthDate': {
        'year': 1971,
        'month': 4,
        'day': 2
        }
        },
        {
        'beneficiaries': [],
        'phoneNumber': '+1 800 1234 568',
        'name': 'Jenny Mack',
        'birthDate': {
        'year': 1951,
        'month': 8,
        'day': 11
        }
        }
        ],
        'signDate': {
        'year': 2021,
        'month': 2,
        'day': 16
        }
        },
        'doc': {
        'dmsUrl': 'http://dms.internal.pbvintage.com/050bca79-5aba-4e32-a34d-9409edcb0a68',
        'login': 'PB\\cnorris',
        'generationDate': {
        'year': 1970,
        'month': 7,
        'day': 20
        }
        }
        }
        }"
    GetTemplatesResponse:
      type: object
      description: "Response type for returning the available templates"
      additionalProperties: false
      properties:
        page:
          $ref: "#/components/schemas/Page"

        contents:
          type: array
          description: "List of template elements."
          items:
            $ref: "#/components/schemas/TemplateDocument"
          maxItems: 32768
    TemplateJobApiResponse:
      type: object
      additionalProperties: false
      properties:
        transactionId:
          type: string
          minLength: 0
          maxLength: 4000
          pattern: "^[a-zA-Z0-9_\\/-]*$"
        message:
          type: string
          minLength: 0
          maxLength: 4000
          pattern: "^[a-zA-Z0-9_/\\]*$"
...
  parameters:
    pageable:
      description: "The data page to be returned."
      in: query
      name: pageable
      required: false
      schema:
        $ref: "#/components/schemas/Pageable"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
security:
  - ApiKeyAuth: []